requirements.txt
fastapi~=0.94.0
pydantic~=1.10.6
structlog~=23.1.0
tortoise-orm~=0.19.3
asyncpg~=0.27.0
